refactor(search): break embeddings, rerank, and types import cycles
Extract EmbeddingProvider into embeddings/contract.ts and RerankProvider
into rerank/contract.ts so implementations depend on the contract, not
the factory. Move resolveConfiguredEmbeddingProvider into
embeddings/configured-provider.ts to cut the provider-resolve -> index
back-edge. Fold SearchSessionFocus, structured-query, evidence-pack,
completeness, and TunedParameters shapes into types.ts, which now has
zero outgoing subsystem imports. Remaining search->tuning->benchmark
triangle is pinned by mcp recall-tools imports and is handled next.
