+++ src/core/search/embeddings/openai-compat.ts
@@ -19,7 +19,7 @@
 
 import { SearchError } from "../types.ts";
 import type { ResolvedEmbeddingConfig } from "../types.ts";
-import type { EmbeddingProvider } from "./provider.ts";
+import type { EmbeddingProvider } from "./contract.ts";
 import {
   AUTH_STATUSES,
   RETRYABLE_STATUSES,
