+++ src/core/search/embeddings/provider.ts
@@ -8,21 +8,7 @@
 
 import { SearchError } from "../types.ts";
 import type { ResolvedEmbeddingConfig } from "../types.ts";
-
-export interface EmbeddingProvider {
-  readonly name: string;
-  readonly model: string;
-  readonly dimension: number | null;
-  embed(texts: ReadonlyArray<string>): Promise<number[][]>;
-  ping(): Promise<{ ok: true; dimension: number } | { ok: false; reason: string }>;
-  /**
-   * Optional read-and-reset of provider-internal retry tally. The
-   * indexer consumes this after each `embed()` to populate
-   * `IndexStats.embeddingsRetries`. Providers that never retry
-   * (NullProvider, MockEmbeddingProvider) leave this undefined.
-   */
-  consumeRetryCount?(): number;
-}
+import type { EmbeddingProvider } from "./contract.ts";
 
 export function makeProvider(config: ResolvedEmbeddingConfig): EmbeddingProvider {
   // Lazy imports to keep the module graph small for users who never
